# smartcity-openapi.yaml
openapi: 3.0.3
info:
  title: SmartCity - Citizen Reporting & Facility Management API
  description: >
    REST API for the SmartCity citizen-reporting platform. Citizens can register,
    submit reports about urban facilities/issues, vote, flag, and comment on
    reports. Admins manage reports, facilities, and notifications through a
    dedicated admin namespace.


    **Authentication:** the API uses a JWT stored in an `httpOnly` cookie named
    `token` (issued on `/login`, on successful Google OAuth, or refreshed on
    `/profile` updates). There is no `Authorization: Bearer` header in this
    implementation — the cookie is sent automatically by browsers, so API
    clients must persist and resend the `token` cookie.


    **CSRF protection:** every state-changing request (`POST`, `PUT`, `PATCH`,
    `DELETE`) — including `/register` and `/login` — is protected by a
    double-submit CSRF cookie. Clients must first perform a `GET` request
    against any server-rendered page to receive the `x-csrf-token` cookie,
    then echo that value back in the `X-CSRF-Token` request header on every
    mutating call. The single exception is `POST /logout`, which is
    registered before the CSRF middleware in the Express stack and therefore
    does not require the header.


    **Rate limiting:** all `/api/*` routes are subject to a global limiter
    (60 req/min for `GET`, 20 req/min for mutating verbs), and `/register`,
...


...
    real API.
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security: []
paths:
...
  /auth/google/callback:
    get:
      tags: [Authentication - OAuth (Browser Flow)]
      summary: Google OAuth callback
      description: >
        Handles Google's redirect back to the app. On success, issues the
        `token` JWT cookie and redirects to `/pages/admin.html` or
        `/pages/menu.html` depending on role (or completes the account-deletion
        re-auth handshake and redirects to `/pages/profile.html`). On failure,
        redirects to `/login.html`.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: OAuth authorization code supplied by Google.
      responses:
        '302':
          description: Redirect to an app page, with the session cookie(s) set on success.
  /auth/google/reauth:
    get:
      tags: [Authentication - OAuth (Browser Flow)]
      summary: Trigger Google re-authentication (for account deletion)
      description: >
        Requires an existing `token` cookie. Sets a short-lived
        `reauth_action` cookie and redirects to Google's consent screen with
        `prompt=consent`. The resulting callback issues a `reauth_token`
        cookie consumed by `DELETE /account`.
      security:
        - cookieAuth: []
      responses:
        '302':
          description: Redirect to Google's consent screen, or to `/login.html` if not authenticated.

  # ============================================================
  # HEALTH
  # ============================================================
...
  /api/facilities:
    get:
      tags: [Facilities (Public)]
      summary: List all facilities
      description: Returns all facilities ordered by name (ascending). Public, no authentication required.
      responses:
        '200':
          description: List of facilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Facility'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          description: Failed to fetch facilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: Gagal mengambil data fasilitas
...
  /api/reports:
    post:
      tags: [Reports (Citizen)]
      summary: Create a new report
      security:
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Report'
        '400':
          description: Missing required fields, or the uploaded image failed buffer validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingFields:
                  value: { message: "Judul dan deskripsi wajib diisi" }
                invalidImage:
                  value: { message: "File tidak valid. Hanya JPEG, PNG, dan WebP yang diizinkan." }
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: Silakan login
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/reports/{reportId}/comments:
    get:
      tags: [Comments]
      summary: List top-level comments (with nested replies) for a report
      description: >
        Requires the caller to be logged in (any role). Returns top-level
        comments ordered oldest-first, each with its `Replies` nested inline.
        Each comment includes a computed `can_edit` flag for the current
        caller. **Implementation note:** internally Express resolves this
        path's parameter as `:id` (merged from the parent router); it is
        renamed to `reportId` here purely for documentation clarity.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportIdForComments'
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/UnauthorizedIsAuth'
        '404':
          $ref: '#/components/responses/ReportNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Comments]
      summary: Add a comment (or reply) to a report
      description: >
        Comments cannot be added to reports with status `hidden`. Supply
        `parent_id` to post a threaded reply to an existing comment on the
        same report.
      security:
        - cookieAuth: []
        - csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ReportIdForComments'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Empty content, or invalid/mismatched `parent_id`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                emptyContent:
                  value: { message: "Komentar tidak boleh kosong" }
                invalidParent:
                  value: { message: "Parent komentar tidak valid" }
        '401':
          $ref: '#/components/responses/UnauthorizedIsAuth'
        '403':
          description: Report does not accept comments (status is `hidden`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: Laporan tidak menerima komentar
        '404':
          $ref: '#/components/responses/ReportNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/admin/reports:
    get:
      tags: [Admin - Reports]
      summary: List reports (admin)
      description: >
        Returns a raw array (not wrapped in `success`/`data`). No query
        validation middleware is attached: any `sort_by` other than
        `vote_count` falls back to `created_at`, and any `order` other than
        exactly `"ASC"` (case-insensitive) falls back to `DESC` — invalid
        values are not rejected with `400`.
      security:
        - cookieAuth: []
      parameters:
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, vote_count]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, resolved, hidden]
        - name: facility_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Matches against title, description, and location_text.
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
            default: 50
          description: Clamped server-side to a maximum of 100.
      responses:
        '200':
          description: Array of report summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminReportListItem'
        '401':
          $ref: '#/components/responses/UnauthorizedIsAdmin'
        '403':
          $ref: '#/components/responses/ForbiddenIsAdmin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/AdminServerError'
...
  /api/admin/facilities:
    get:
      tags: [Admin - Facilities]
      summary: List all facilities (admin)
      description: Returns a raw array (not wrapped in `success`/`data`), includes `image_public_id`.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facility'
        '401':
          $ref: '#/components/responses/UnauthorizedIsAdmin'
        '403':
          $ref: '#/components/responses/ForbiddenIsAdmin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/AdminServerError'

    post:
      tags: [Admin - Facilities]
      summary: Create a facility (admin)
      security:
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FacilityCreateRequest'
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: Facility created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
        '400':
          description: Uploaded image failed buffer validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "File tidak valid. Hanya JPEG, PNG, dan WebP yang diizinkan."
        '401':
          $ref: '#/components/responses/UnauthorizedIsAdmin'
        '403':
          $ref: '#/components/responses/ForbiddenIsAdmin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/AdminServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: >
        JWT issued on login (local or Google) and decoded on every request by
        server middleware. Contains `id`, `name`, `email`, `role`, `phone`,
        `address`, `provider`. There is no Bearer-token alternative in this
        implementation.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >
        Double-submit CSRF token. Obtain the current value from the
        `x-csrf-token` cookie (set by the server when rendering any HTML
        page, and exposed in a `<meta name="csrf-token">` tag) and echo it
        back in this header on every `POST`/`PUT`/`PATCH`/`DELETE` request.
        Form submissions may alternatively send it as a `_csrf` body field.
...
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            auth:
              summary: On /register, /login, /profile (5 / 15 min)
              value: { success: false, message: "Terlalu banyak percobaan, coba lagi nanti." }
            read:
              summary: On GET /api/* (60 / min)
              value: { success: false, message: "Terlalu banyak permintaan, tunggu sebentar." }
            write:
              summary: On mutating /api/* (20 / min)
              value: { success: false, message: "Terlalu banyak aksi, silakan tunggu sebentar." }
...
    AdminServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleError'
...
  schemas:
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
      required: [message]
...